+++ app/models/work_package.rb
@@ -141,6 +141,8 @@ class WorkPackage < ActiveRecord::Base
   acts_as_attachable :after_add => :attachments_changed,
                      :after_remove => :attachments_changed
 
+  after_validation :set_attachments_error_details, if: lambda {|work_package| work_package.errors.messages.has_key? :attachments}
+
   # Mapping attributes, that are passed in as id's onto their respective associations
   # (eg. type=4711 onto type=Type.find(4711))
   include AssociationsMapper
@@ -284,8 +286,10 @@ class WorkPackage < ActiveRecord::Base
   # ACTS AS ATTACHABLE
   # Callback on attachment deletion
   def attachments_changed(obj)
-    add_journal
-    save!
+    unless new_record?
+      add_journal
+      save
+    end
   end
 
   # ACTS AS JOURNALIZED
@@ -407,13 +411,9 @@ class WorkPackage < ActiveRecord::Base
 
     update_by(user, attributes)
 
-    if save
-      # as attach_files always saves an attachment right away
-      # it is not possible to stage attaching and check for
-      # valid. If this would be possible, we could check
-      # for this along with update_attributes
-      attachments = Attachment.attach_files(self, raw_attachments)
-    end
+    attach_files(raw_attachments)
+
+    save
   end
 
   def update_by(user, attributes)
@@ -1001,4 +1001,12 @@ class WorkPackage < ActiveRecord::Base
   end
   # <<< issues.rb <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
 
+  def set_attachments_error_details
+    # Remark: the pseudo loop is already refactored in the next pull request
+    self.attachments.each do |attachment|
+      next if attachment.valid?
+      errors.messages[:attachments].first << " - #{attachment.errors.full_messages.first}"
+      break
+    end
+  end
 end